refactor: consolidate reshaping functions under the reshape package (#1194)
* consolidate reshaping functions under the reshape package

* format code

* fix import

* fix import

* lint code
